implemented Petrovic et al (2010) and wrote a program to test it on the TDT5 dataset
git-svn-id: http://svn.basistech.net/engineering/u/jwp/lshj@13598 81b6c55f-58c9-4729-b236-40b2a262f43e

+++ src/com/basistech/lsh/Projection.java
@@ -2,12 +2,12 @@ package com.basistech.lsh;
 
 public class Projection {
     private int size;
-    private int[] rep;
+    private double[] rep;
 
     public Projection(int size, Sampler sampler) {
         assert(size > 0);
         this.size = size;
-        rep = new int[size];
+        rep = new double[size];
         for (int i = 0; i < size; ++i) {
             rep[i] = sampler.draw();
         }
